--- src/Data display/Badge/Badge.stories.js
@@ -1,13 +1,0 @@
-import React from "react";
-import Badge from ".";
-
-export default {
-  title: "Data display/Badge",
-  component: Badge
-};
-
-const Template = (args) => (
-  <Badge />
-);
-
-export const Default = Template.bind({});
