+++ src/Data display/Chip/Chip.stories.mdx
@@ -1,11 +1,11 @@
 import { Meta, Story, Canvas, Props } from "@storybook/addon-docs/blocks";
 import { ArgsTable, Stories } from "@storybook/addon-docs";
 import Chip from "./index";
-import Avatar from "../Avatar";
+import Avatar from "../../General/Avatar";
 import logo from '../../_utils/images/Image.png'
 
 
-<Meta title="General/Chip" component={Chip} />
+<Meta title="Data display/Chip" component={Chip} />
 
 export const TemplateDefault = (args) => <Chip {...args} />;
